SPM: Move xlat-related functions to separate file
This is done in order to make it easier to read the file spm_main.c.
