+++ services/std_svc/spm/spm.mk
@@ -15,7 +15,8 @@ SPM_SOURCES	:=	$(addprefix services/std_svc/spm/,	\
 			${ARCH}/spm_helpers.S			\
 			${ARCH}/spm_shim_exceptions.S		\
 			spm_main.c				\
-			sp_setup.c)
+			sp_setup.c				\
+			sp_xlat.c)
 
 
 # Let the top-level Makefile know that we intend to include a BL32 image
